[WIP] Refactored the agent's profiles
Instead of having a '_load_profile' function in the 'scenarios' package (and no 'save' function), the notion of profile was refactored in a Profile class, in the 'util' package.
This makes Profile safer (by controlling how to save and load), and easier to use (since we now have attributes instead of accessing a dictionary).