[WIP] Refactored the available energy
We previously used a simple function 'compute_available_energy'.
However, we also need to know the "bounds" for this amount of available energy, which depends on the actual function used.
Instead of using 2 different functions, knowing that one of them depends on the other, this commit adds a new abstract class, 'EnergyGenerator', which contains these 2 functions.
Several implementations are already provided: 'RandomEnergyGenerator', 'ScarceEnergyGenerator', 'GenerousEnergyGenerator', and 'RealisticEnergyGenerator'.

Also, the 'Observations' now uses these bounds to scale the available energy to (0,1).