Added the TestModels test file
This file tests that the simulator works correctly, as a whole, by using two different models: a random model (no real "learning" algorithm, just picking a random action each step), and the QSOM model.

This file can also be used as inspiration to implement new models (provides the basic, "boilerplate" code to instantiate the simulator and use a model on it).