Removed the scenarios package
The package is no longer up-to-date with the rest of the `smartgrid` code, and contained too much hard-coded instructions for instantiating envs.
In addition, it was not very useful as-is: users will want to create their own scenarios.
In this regard, it can still be used as a source of inspiration.