Fix inherited members in docs
Sphinx (autosummary) raises errors when documenting members (especially attributes) of a parent class that do not appear in the child class, e.g., the `Reward`'s `name` in this codebase.

This is a known bug: https://github.com/sphinx-doc/sphinx/issues/9884

Added at least the attribute(s) declaration (type) in each of the child classes (for `Reward`, `DataConversion`, and `EnergyGenerator`), which seems sufficient to make Sphinx acknowledge that the attribute exists, and pull up the documentation from the parent class.