Load in DataConversion now returns the Profile
The `load` method of `DataConversion` now returns the profile, in addition to memorizing it in its `profiles` attribute.
This simplifies testing and debugging, by directly using the requested profile in a single expression `profile = converter.load(...)` rather than 2 `converter.load(...); profile = converter.profiles[...]`.
The profile remains available in the `profiles` attribute, so this change is backward-compatible with previous code.