Fixed the agents' needs in TestWorld
We check if the computed comfort is coherent with the comfort function, based on the agents' consumption and needs.
However, the previous code used the need *after* the `world.step()` call, which updates the agents, including their needs. In other words, we compared `comfort(consumption at t, need at t)`, with `comfort(consumption at t, need at t+1)`.
The new code memorizes the agents' needs to solve this bug.

Additionally, added a for loop to test for several steps instead of a single one. The `self.subTest()` call gives more contextual information in case of failure(s).