Updated requirements for Python3.7
Latest versions of NumPy have dropped support for Python3.7 (in their metadata), which means pip cannot find a suitable version.
Thus, lowered the minimum version to '1.19.0', which should support Python3.7.
As this is a minimum version, users with a more recent Python version should still get a recent NumPy version.
Python3.7 is not officially supported any more, but in case it is still used, we prefer to support it.