Fixed images links in docs
The "architecture" image was supposed to be clickable, to be easily seen (large image, not very readable when in-line).
However, the `target` was not correct: Sphinx actually requires the path relative to the built documentation (not the source). In addition, Sphinx moves the image to a `_images/` folder, in the build root, when building. To make sure we can access it, we use relative paths; this ensures the documentation will work even when deployed to GitHub Pages, where it is in a sub-folder (and thus `/_images/...` would not work).