Add the argumentation-based reward functions
Added the following new classes that derive from `Reward`:
* Inclusiveness
* EnvironmentalSustainability
* Affordability
* SupplySecurity

Because the various moral values are implemented in different classes,
they can be used either individually or at the same time.
However, the `parse_situation` helper must be called for each moral value,
which implies a small (time) cost.

We also include a custom error message in the module's
`__init__.py`, because AJAR is an optional dependency
(so some users may not have it), and because it is only
accessible through a git+https link, not through PyPi
(i.e., `pip install ajar` will not work, so we must tell
our users to type instead `pip install git+https://...`).