Refactored Local and Global Observations to use dataclasses instead of namedtuples
The classes for the Observations used `namedtuple`, which was great to ensure
they are immutable, and that users can access their fields by names, e.g.,
`obs.comfort` (instead of having to find their index in a given list).

However, namedtuples bring 2 major disadvantages:
* they cannot be inherited (and thus extended) easily;
* only the fields' names are specified, not their types.

Instead, `dataclasses` are a more recent and better tool to create such classes
that are immutable, which fields can be accessed by names, to which we can add
functions, and which can be extended through inheritance.

This commit thus refactors those classes to use `dataclasses` instead; the
usage should be quite similar, with the following additions:
* The `fields` method can be used to get the fields' names (whereas it
  previously required defining the list of names somewhere or using private
  methods).
* The `asdict` method can be used to represent the observation as a dictionary,
  when a dataclass does not do the job.
* The `__array__` magic method simplifies the transformation into a NumPy
  ndarray: an observation can be passed to `np.ndarray(obs)` to efficiently
  (and easily) obtain such arrays. This is particularly useful, since learning
  and decision-making algorithms typically use ndarrays. We can thus pass
  dataclasses around, knowing that they will be converted automatically to
  ndarrays when necessary.

The `Observation` class is temporarily changed to avoid raising exceptions
(because it can no longer access the `GlobalObservation` and `LocalObservation`
fields), but will be deeply refactored.