Fixed Quick usage instructions
The quick usage example code would provoke a SyntaxError when used in a Python REPL (but, curiously, not in an IPython REPL), as per #1.
Added a blank line to solve this syntax error.
Additionally, added the `max_step=10` argument, to limit the number of simulated time steps, and avoid hanging for ~ 1 hour (very confusing for new users, especially since there is no output on stdout to show that something is happening).
Finally, added a small paragraph to explain what the Quick usage examples does.