Update quick usage example to print rewards
The quick usage previously did not print *anything* (silent run). This was made
voluntarily, to avoid "cluttering" the standard output, but has the clear
disadvantage of not showing that "something is happening". It is a problem,
especially when using many time steps, which can be very long (e.g., several
minutes).

Instead, the quick usage was updated with 2 objectives in mind (in line with issue
#3): a) show that something is happening, and b) show the user intuitively how
algorithms could be compared.

Any basic message could be used to support objective a) (e.g., "Time step X"),
but, to support objective b) as well, we have chosen to print the agents' rewards,
at each time step.

The `SmartGrid` class itself is not modified, and still does not print anything by
itself, so that we can control its output through the interaction loop. Users
choose by themselves what to print. Verbosity parameters could have been used,
yet there is so much possible information to print that those verbosity parameters
would quickly become very tedious.

Also added a few links to the documentation in the Readme, so that users have some
hints on what to do after running the Quick Usage example.