Added list and description of pre-defined components in documentation
The API page was a bit empty, with only the class diagram and the links to
the top-level packages (`smartgrid` and `algorithms`).
Added below, as per issue #3, a list of pre-defined components (agents
profiles and DataConversion to create them; energy generators; reward
functions; learning algorithms), so that the API can be a bit simpler to
explore, and users see at a glance what they can leverage for their own
experiments.