Fix bug when resetting Env
In order to re-use an environment, we are supposed to use the reset method.
This method is responsible for re-initializing all components so that they are in a "clean state". This includes the agents, of course, but also the ObservationManager (particularly the GlobalObservation), because it uses a counter to memorize the last computed step!
This method was written, but never called, which produced a bug in a single specific situation: when launching a new simulation after a 1-step simulation. Indeed, the last computed step would be in this case `t=0`, the first step, which corresponds to the "new first step" of the new simulation, and thus returns the observations from the previous simulation.
Any other setup with more than 1 step would avoid this bug, because `t>0` would not match the "new first step".

Just for good measure, we now also reinitialize the `computed` variable to `None`. Reinitializing the counter should suffice, but it is clearer that way (and also less confusing for a dev debugging several simulations).

On a side note, using class methods and attributes in Observations (Global especially) facilitated this bug, because re-creating the Env does not reset the "static values". This also means that we cannot run several simulations in parallel, because of the implicit data sharing implied by the class attribute of GlobalObservation.