Refactor EnergyGenerator to allow setting local random seed
Instead of setting the random seed globally, we want to parameterize each component individually. This avoids bugs (side-effects) and also will allow to run several simulations in parallel.

The EnergyGenerator component (and particularly the RandomGenerator derived class) relied on `random.randint` to generate random values, which uses the global seed.
Instead, added a new attribute with a NumPy Generator instance, and a method to set this instance to a given Generator (which effectively sets the seed).

This commit also fixes the related test, which relied on globally setting the `random.seed()` and did not work any more.