Setting the random seed correctly in `env.reset()`
To ensure reproducibility, the Gymnasium `env.reset()` generates a NumPy random Generator when the seed is set.
However, this variable was not accessible to our components; refactored the World class to receive it as a parameter of its `world.reset()` method, and to pass it to the components, especially the EnergyGenerator.

To avoid breaking the API (forcing users to set this parameter whenever they use
`world.reset()`), this parameter has `None` as its default value. When the value
is `None`, a default random Generator is created, by using the recommended NumPy
`np.random.default_rng()` (without a seed).