Added sphinx-multiversion to generate all docs
This adds the ability to generate the docs for multiple versions,
automatically retrieved from Git tags and branches.
Docs are built separately (but from the same conf.py file!!), and
each version gets its own folder.

The `conf.py` file specifies which branches and tags are taken into
account. (We explicitly remove branch `paper`, tag `v1.1.0-joss-paper`,
and any branch beginning with `wip/` for now)

We also modified the sidebar to include a custom version selector,
allowing users to switch versions when reading the docs.
The theme we use (Furo) is currently trying to add official support
to multiple versions (either with sphinx-multiversion or any similar
tool), which should simplify this selector in the future.

Note: this version selector is NOT internationalized, because we wanted
custom headings (e.g., "Released versions" instead of just "Versions",
which can be obtained with `_('Version')` to support internationalization.
Because our doc is only available in english, it should not be a problem...