Added multiversion rule to Makefile and redirection
To simplify building the (multiple) docs, added a `multiversion` rule
to the Makefile, similar to the previous rules (e.g., `make html`).

In addition, this rule will also create a fake `index.html` file in the
build root folder, which will simply redirect the client to one of the versions.
Because each version gets its own sub-folder, and `sphinx-multiversion` does not
provide any "default" landing page, clients would not be able to access the
documentation without this redirection (they would get a 404 Not Found page,
and can only access the sub-folders if they know their names!).

This fake `index.html` must know to which path it should redirect clients.
Hard-coding a path would work, but if the folder is not found (for some reason,
e.g., the version not being built), then the redirection would end up in a 404
again.
Instead, in the Makefile, after building all versions, we automatically search
for the latest tagged release (e.g., `v1.0.0`; we ignore branches because they
are unstable versions).
An advantage of this approach is that the found "latest" folder must exist,
because we list existing folders to identify it. If no such "latest" version
can be found, we abort the build process to avoid pushing a malformed doc and
presenting users with a 404 when they reach the website.
To simplify, the template `redirect_to_version.html` is rendered with `sed`
to put the correct path when copying it to the build folder. (This avoids
relying on more complex templating systems, such as Jinja2)

Note: it could be useful to have a variable that specifies the desired version;
if the variable is set, we check that the folder exists. If the folder does not
exist, or the variable is not set, we resort to finding the latest version
automatically.
However, Make does not allow easily creating variables within a target body
(which is also why all these steps are defined as a single shell command, split
by `;/` to use newlines while convincing Make that they are part of the same
command).