Update GitHub workflow to build multiversion docs
The workflow is now triggered by a push to any branch (instead of just
`master` and `docs` previously), except `wip/*` branches that are ignored.

It uses the Make `multiversion` rule, which builds the docs into `html-mv`.

We also change the checkout step to specify `fetch-depth: 0`.
If we do not specify it, only the triggered branch is fetched, which
prevents sphinx-multiversion from building all docs (branches and tags).