Fix hatch build
PyPi does not allow using direct references (URL) to specify
dependencies. We must thus remove `ajar` from the list
of (optional) dependencies in the `pyproject.toml`.

AJAR may remain a dependency in the `requirements.txt` file,
which means that we can still use it in CI/CD workflows.

Instead, this commit changes the `Readme.md` to instruct
users on how to install ethical-smart-grid and optionally
ajar.