Improve the Observation dataclass
Force the qualname (fully qualified name) to be `'Observation'` instead of the longer
(but more accurate) `_create_observation_type.<locals>._Observation`. This is
especially important when people want to get a string representation of an Observation
(through the `str` or `repr` methods), because dataclasses return a string in the
following format: `qualname(field1=value1, field2=value2, ...)`.
When qualname is long and difficult to read (`<locals>`, etc.), this might confuse
users.

Another approach, which would not require to "lie" on the qualname, would be to
override the `__repr__` method, to reproduce a similar behaviour to the default
dataclass but hardcoding the first part to be `Observation` instead of the qualname.
It is a bit more complicated however, as it requires re-parsing the fields.