Refactor reward wrappers (aggregators) for PettingZoo
Our custom wrappers used the Gymnasium API, which is not compatible with
PettingZoo. In addition, PettingZoo only has wrappers for their AEC envs
(turn-by-turn, not all agents in parallel). It would hinder performances
to transform our environment to AEC, wrap it, then back to Parallel.
Instead, we have made our `RewardAggregator` base class the equivalent
of the `RewardWrapper` from Gymnasium: it provides access to the wrapped
class seamlessly and allows to intercept the rewards through its
`reward` method.

All our reward aggregators now return dicts instead of lists, to follow
the PettingZoo API.

It would have been overkill to allow intercepting other values as well
(e.g., obs, infos, ...), because we have no need for it now. It would be
fairly easy to do in the future if it becomes necessary (provide other
methods, and call them in `step` similarly to how we call `reward`).

Because we inherit from `SmartGrid` to make the wrapper act as a drop-in replacement, the magic method `__getattr__` does not "intercept" method calls if they are found in the inheritance graph. Instead, we have to use `__getattribute__` which intercepts *all* accesses and is a bit more complicated to use.
All attributes or methods that must be accessed directly in the wrapper (instead of the wrapped env) *MUST* start with an underscore (`_`) or be listed as an exception in `__getattribute__`. Otherwise, the code will crash or loop infinitely.