Improve the `find_profile_data` function for Python3.9+
The `importlib.resources` lib that we use in this function has changed its API. Our function would still work but raise a deprecation warning. There is a risk that the functions we require would be deleted at some point; this commit makes our function future-proof (at least for the next versions).
We try to import the most recent functions; if it does not work, we assume we are on an older version and resort to the previous API (for which there will be no deprecation warning at this point).

Python 3.7 is no longer supported by the Python community; Python 3.8 will reach end-of-life on 2024-10. At some point, our function could drop this "backup" import for older versions; but for now, they may still be used.