Add new method to get optimal actions in a Model.
The existing `forward()` method of Models produces actions from observations, and is meant to be used in the in training phase. This means that most Models will include notions of *exploitation-exploration* in this method, e.g., trying an action with a low value to see if it yields a better reward than expected.

When testing / validating a Model, we might prefer to focus on *exploitation*, i.e., returning the actions that the Model considers the best for a given state. This is the goal of the new `get_optimal_actions()` method.

By default, it returns the same actions as `forward`, to ensure that all Models will be usable, even if they do not override it.
The QSOM model overrides this method to bypass the Boltzmann exploration and random perturbation.